+++ app/examples/AndroidManifest.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.pubnub.crc.examples">
+    package="resourcecenterdemo">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
-        android:name="com.pubnub.crc.examples.App"
+        android:name=".App"
         android:allowBackup="false"
         android:appComponentFactory="androidx.core.app.CoreComponentFactory"
         android:icon="@mipmap/ic_launcher"
@@ -18,7 +18,7 @@
         tools:replace="android:appComponentFactory"
         tools:targetApi="p">
         <activity
-            android:name="com.pubnub.crc.examples.SplashActivity"
+            android:name=".SplashActivity"
             android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -27,7 +27,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.pubnub.crc.examples.MainActivity"
+            android:name=".MainActivity"
             android:theme="@style/AppTheme" />
     </application>
