--- app/examples/com/pubnub/crc/examples/util/PNFragment.java
@@ -1,9 +1,0 @@
-package com.pubnub.crc.examples.util;
-
-import com.pubnub.api.callbacks.SubscribeCallback;
-
-public interface PNFragment {
-
-    SubscribeCallback provideListener();
-
-}
