--- app/examples/com/pubnub/crc/examples/util/PNGlideModule.java
@@ -1,9 +1,0 @@
-package com.pubnub.crc.examples.util;
-
-import com.bumptech.glide.annotation.GlideModule;
-import com.bumptech.glide.module.AppGlideModule;
-
-@GlideModule
-public class PNGlideModule extends AppGlideModule {
-
-}
