--- app/examples/com/pubnub/crc/examples/util/ParentActivityImpl.java
@@ -1,15 +1,0 @@
-package com.pubnub.crc.examples.util;
-
-import com.pubnub.api.PubNub;
-
-import androidx.fragment.app.Fragment;
-
-public interface ParentActivityImpl {
-
-    PubNub getPubNub();
-
-    void setTitle(String title);
-
-    void addFragment(Fragment fragment);
-
-}
