+++ app/examples/resourcecenterdemo/pubnub/PubNubListener.java
@@ -1,0 +1,25 @@
+package resourcecenterdemo.pubnub;
+
+import com.pubnub.api.PubNub;
+import com.pubnub.api.callbacks.SubscribeCallback;
+import com.pubnub.api.models.consumer.PNStatus;
+import com.pubnub.api.models.consumer.pubsub.PNMessageResult;
+import com.pubnub.api.models.consumer.pubsub.PNPresenceEventResult;
+
+public class PubNubListener extends SubscribeCallback {
+
+    @Override
+    public void status(PubNub pubnub, PNStatus status) {
+
+    }
+
+    @Override
+    public void message(PubNub pubnub, PNMessageResult message) {
+
+    }
+
+    @Override
+    public void presence(PubNub pubnub, PNPresenceEventResult presence) {
+
+    }
+}
