+++ app/examples/resourcecenterdemo/util/PNFragment.java
@@ -1,0 +1,9 @@
+package resourcecenterdemo.util;
+
+import com.pubnub.api.callbacks.SubscribeCallback;
+
+public interface PNFragment {
+
+    SubscribeCallback provideListener();
+
+}
