+++ app/examples/resourcecenterdemo/util/PNGlideModule.java
@@ -1,0 +1,9 @@
+package resourcecenterdemo.util;
+
+import com.bumptech.glide.annotation.GlideModule;
+import com.bumptech.glide.module.AppGlideModule;
+
+@GlideModule
+public class PNGlideModule extends AppGlideModule {
+
+}
