+++ app/examples/resourcecenterdemo/util/ParentActivityImpl.java
@@ -1,0 +1,15 @@
+package resourcecenterdemo.util;
+
+import com.pubnub.api.PubNub;
+
+import androidx.fragment.app.Fragment;
+
+public interface ParentActivityImpl {
+
+    PubNub getPubNub();
+
+    void setTitle(String title);
+
+    void addFragment(Fragment fragment);
+
+}
