+++ config/checkstyle/checkstyle.xml
@@ -38,7 +38,7 @@
         <property name="basedir" value="${basedir}"/>
     -->
 
-    <property name="severity" value="warning" />
+    <property name="severity" value="error" />
 
     <property name="fileExtensions" value="java, properties, xml" />
 
@@ -151,7 +151,7 @@
         <!-- Checks for blocks. You know, those {}'s         -->
         <!-- See http://checkstyle.sf.net/config_blocks.html -->
         <module name="AvoidNestedBlocks" />
-        <module name="EmptyBlock" />
+        <!--<module name="EmptyBlock" />-->
         <module name="LeftCurly" />
         <module name="NeedBraces" />
         <module name="RightCurly" />
@@ -161,10 +161,14 @@
         <!--<module name="AvoidInlineConditionals"/> -->
         <module name="EmptyStatement" />
         <module name="EqualsHashCode" />
-        <module name="HiddenField" />
+        <module name="HiddenField">
+            <property name="ignoreFormat" value="pubNub" />
+        </module>
         <module name="IllegalInstantiation" />
         <module name="InnerAssignment" />
-        <module name="MagicNumber" />
+        <module name="MagicNumber">
+            <property name="ignoreFieldDeclaration" value="true" />
+        </module>
         <module name="MissingSwitchDefault" />
         <module name="SimplifyBooleanExpression" />
         <module name="SimplifyBooleanReturn" />
@@ -174,7 +178,9 @@
         <!-- <module name="DesignForExtension"/> -->
         <module name="HideUtilityClassConstructor" />
         <module name="InterfaceIsType" />
-        <module name="VisibilityModifier" />
+        <module name="VisibilityModifier">
+            <property name="packageAllowed" value="true" />
+        </module>
 
         <!-- Miscellaneous other checks.                   -->
         <!-- See http://checkstyle.sf.net/config_misc.html -->
